Move Omnidesk duplicates remover to a common file
