+++ src/populate.ts
@@ -4,28 +4,7 @@ dotenv.config();
 import type { InferCreationAttributes } from "sequelize";
 import { fetchAllOmnideskUsers } from "./omnidesk/api";
 import { User } from "./db";
-import { OmnideskUser } from "./omnidesk/types";
-
-function removeDuplicates(users: OmnideskUser[]): OmnideskUser[] {
-  const userNameCache: Record<string, number> = {};
-  const tgIdCache: Record<string, number> = {};
-  return users.filter((user, i, users) => {
-    if (user.user_screen_name in userNameCache) {
-      const idx = userNameCache[user.user_screen_name];
-      console.warn(`Duplicate telegram alias found:`, user, users[idx]);
-      return false;
-    }
-    userNameCache[user.user_screen_name] = i;
-
-    if (user.telegram_id in tgIdCache) {
-      const idx = tgIdCache[user.telegram_id];
-      console.warn(`Duplicate telegram chat_id found:`, user, users[idx]);
-      return false;
-    }
-    tgIdCache[user.telegram_id] = i;
-    return true;
-  });
-}
+import { removeDuplicates } from "./omnidesk/helpers";
 
 async function main() {
   await User.sync({ force: true });
@@ -39,6 +18,8 @@ async function main() {
       telegramChatId: Number(user.telegram_id),
       registrationLastNotified: null,
       visaLastNotified: null,
+      registrationExpiration: null,
+      visaExpiration: null,
     }));
 
   await User.bulkCreate(dbUsers);
